chore: Move wrapInTableView into its own file
