+++ wire-ios/Wire-iOS Tests/Utils/ZMSnapshotTestCase.swift
@@ -21,52 +21,6 @@ import UIKit
 @testable import Wire
 import WireCommonComponents
 
-extension UITableViewCell: UITableViewDelegate, UITableViewDataSource {
-    func wrapInTableView() -> UITableView {
-        let tableView = UITableView(frame: self.bounds, style: .plain)
-
-        tableView.delegate = self
-        tableView.dataSource = self
-        tableView.backgroundColor = .clear
-        tableView.separatorStyle = .none
-        tableView.rowHeight = UITableView.automaticDimension
-        tableView.layoutMargins = self.layoutMargins
-
-        let size = self.systemLayoutSizeFitting(CGSize(width: bounds.width, height: 0.0), withHorizontalFittingPriority: .required, verticalFittingPriority: .fittingSizeLevel)
-        self.layoutSubviews()
-
-        self.bounds = CGRect(x: 0.0, y: 0.0, width: size.width, height: size.height)
-        self.contentView.bounds = self.bounds
-
-        tableView.reloadData()
-        tableView.bounds = self.bounds
-        tableView.layoutIfNeeded()
-
-        NSLayoutConstraint.activate([
-            tableView.heightAnchor.constraint(equalToConstant: size.height)
-        ])
-
-        self.layoutSubviews()
-        return tableView
-    }
-
-    public func numberOfSections(in tableView: UITableView) -> Int {
-        return 1
-    }
-
-    public func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return 1
-    }
-
-    public func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return self.bounds.size.height
-    }
-
-    public func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        return self
-    }
-}
-
 class ZMSnapshotTestCase: FBSnapshotTestCase {
 
     typealias ConfigurationWithDeviceType = (_ view: UIView, _ isPad: Bool) -> Void
